move fetch-labor-rights to utils
